requirements.txt
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
pymatgen==2024.8.9
mace-torch==0.3.8
torch==2.2.2
matgl==1.0.0
maml==2024.6.13
atomate2==0.0.14

